refactor: Update backend imports in experiment and measurement modules
